# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "utopia API",
        "description": "API documentation for utopia ",
        "contact": {
            "email": "info@utopia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gestion.utopia.madrid/api/",
            "description": "utopia API"
        }
    ],
    "paths": {
...
        "/entities/{id}": {
            "get": {
                "tags": [
                    "entidades"
                ],
                "summary": "Mostrar entidad detalle",
                "operationId": "fa15041858ccfde07b3398422de1005e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta de entidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "entidades"
                ],
                "summary": "Modificar entidad",
                "operationId": "e656bbb370b39819b530f1de81f02512",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar entidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "put entidad",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "lon": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "idcategoria": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "idsector": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "informacion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "urlweb": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "entidades"
                ],
                "summary": "Eliminar entidad",
                "operationId": "b72acbd76972372a06ec10334f089864",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar entidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}